+++ src/main/java/org/internship/model/Supervisor.java
@@ -10,16 +10,16 @@ import java.util.List;
 public class Supervisor extends User {
   @Id
   @GeneratedValue
-  private Long id;
+  private int id;
   private String cpf;
   @OneToMany
   private List<Internship> internships;
 
-  public Long getId() {
+  public int getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(int id) {
     this.id = id;
   }
